requirements.txt
absl-py==2.1.0
accelerate==1.3.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
diffusers==0.32.2
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
keras==3.8.0
kombu==5.4.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
